backend/requirements.txt
fastapi 
uvicorn
psycopg2-binary
sqlalchemy
alembic 
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
httpx
redis
celery